[FEAT] Add Calendar View Link
- Add Link from Calendar View to Daily View
